% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Launch the Snply Shiny App}
\description{
  This function launches the Snply Shiny app, which allows users to upload SNP data files and analyze the data.
}
\usage{
  run_app()
}
\value{
  This function does not return any value. It runs a Shiny application.
}
\details{
  The Snply Shiny app provides a user interface for:
  - Uploading SNP files in various formats (e.g., .txt, .csv, .tsv, .vcf).
  - Displaying the SNP data in an interactive table.
  - Plotting heterozygosity vs. homozygosity of the SNP data.

  The file size limit is 50MB.
}
\examples{
  # Launch the Snply app
  run_app()
}


