% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\name{calculate_neanderthal_alleles}
\alias{calculate_neanderthal_alleles}
\title{Calculate Neanderthal Alleles Summary}
\usage{
calculate_neanderthal_alleles(user_file)
}
\description{
Calculate Neanderthal Alleles Summary
}
