% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplogroups_function.R
\name{confidence_level}
\alias{confidence_level}
\title{Confidence Level from Match Count}
\usage{
confidence_level(n)
}
\arguments{
\item{n}{SNP match count}
}
\value{
"High", "Medium", "Low", "None", or "Unknown"
}
\description{
Confidence Level from Match Count
}
