% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{draw_chromosome_silhouette}
\alias{draw_chromosome_silhouette}
\title{Draw chromosome silhouettes (scaled, sorted 1–22, X, Y)}
\usage{
draw_chromosome_silhouette()
}
\value{
ggplot
}
\description{
Draw chromosome silhouettes (scaled, sorted 1–22, X, Y)
}
