% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplogroups_function.R
\name{load_haplogroup_references}
\alias{load_haplogroup_references}
\title{Load Haplogroup Reference Data}
\usage{
load_haplogroup_references()
}
\value{
A list with mt_ref, y_clean, haplo_map
}
\description{
Load Haplogroup Reference Data
}
