% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_mc.R
\name{paint_mc}
\alias{paint_mc}
\title{Monte-Carlo ancestry painting}
\usage{
paint_mc(user_snps, ref_freq, block_n = 50, iter = 150, thresh = 0.8)
}
\arguments{
\item{user_snps}{data.table with columns rsid, chromosome, position, genotype}

\item{ref_freq}{data.table from load_aisnp_ref(), with columns SNP,CHR,A1,CEU,CHB,GIH,JPT,YRI}

\item{block_n}{SNPs per block}

\item{iter}{number of MC iterations}

\item{thresh}{mono-ancestry vote threshold (0–1)}
}
\value{
data.table: CHR, start, end, hap1, hap2
}
\description{
Monte-Carlo ancestry painting
}
