% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{paint_snps}
\alias{paint_snps}
\title{Paint introgressed SNPs on top of a chromosome silhouette}
\usage{
paint_snps(base_plot, snps, tick_shape = c("segment", "rect"))
}
\arguments{
\item{base_plot}{Output of draw_chromosome_silhouette()}

\item{snps}{Data.frame with columns chromosome, position (bp), archaic_count}

\item{tick_shape}{"segment" (default) or "rect"}
}
\value{
ggplot
}
\description{
Paint introgressed SNPs on top of a chromosome silhouette
}
