% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_ancestry.R
\name{plot_all_chromosomes}
\alias{plot_all_chromosomes}
\title{Plot genome-wide ancestry painting (all chromosomes)}
\usage{
plot_all_chromosomes(all_blocks, pop_cols)
}
\arguments{
\item{all_blocks}{data.table with columns CHR, block, start, end, hap1, hap2}

\item{pop_cols}{named vector of colors for each ancestry}
}
\value{
ggplot
}
\description{
Plot genome-wide ancestry painting (all chromosomes)
}
