% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplogroups_function.R
\name{plot_haplogroup_location}
\alias{plot_haplogroup_location}
\title{Plot Haplogroup Dispersion Map}
\usage{
plot_haplogroup_location(haplogroup, haplo_map)
}
\arguments{
\item{haplogroup}{Haplogroup string}

\item{haplo_map}{Reference map of haplogroup locations}
}
\value{
A leaflet map object
}
\description{
Plot Haplogroup Dispersion Map
}
