% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplogroups_function.R
\name{snply_detect_maternal_haplogroup}
\alias{snply_detect_maternal_haplogroup}
\title{Detect Maternal Haplogroup}
\usage{
snply_detect_maternal_haplogroup(user_mt, mt_ref)
}
\arguments{
\item{user_mt}{Data frame of mitochondrial SNPs}

\item{mt_ref}{Reference data for mtDNA haplogroups}
}
\value{
A list with best match haplogroup and match count
}
\description{
Detect Maternal Haplogroup
}
