% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplogroups_function.R
\name{snply_detect_paternal_haplogroup}
\alias{snply_detect_paternal_haplogroup}
\title{Detect Paternal Haplogroup by Y Position}
\usage{
snply_detect_paternal_haplogroup(user_y, y_ref, top_n = 1)
}
\arguments{
\item{user_y}{Data frame of Y chromosome SNPs}

\item{y_ref}{Cleaned reference of Y-SNPs}

\item{top_n}{Number of top matches to return}
}
\value{
A list with haplogroup, match count, and cleaned display name
}
\description{
Detect Paternal Haplogroup by Y Position
}
