% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\name{calculate_neanderthal_alleles}
\alias{calculate_neanderthal_alleles}
\title{Calculate Neanderthal Alleles Summary}
\usage{
calculate_neanderthal_alleles(user_file)
}
\arguments{
\item{user_file}{Path to the user's 23andMe genotype file.}
}
\value{
A list with a summary string, count of allele copies, percentage of allele copies, and a phenotype table.
}
\description{
Orchestrates the reading, merging, and computation of Neanderthal allele statistics.
}
