% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\name{compute_allele_statistics}
\alias{compute_allele_statistics}
\title{Compute Neanderthal Allele Statistics}
\usage{
compute_allele_statistics(merged_df)
}
\description{
Compute Neanderthal Allele Statistics
}
