% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neanderthal.R
\name{read_user_data}
\alias{read_user_data}
\title{Read User Data (e.g., 23andMe genotype file)
Reads the user's genotype file, accommodating headers that may
or may not be commented out with '#'. Keeps rsid, chromosome, position, genotype.}
\usage{
read_user_data(filepath)
}
\arguments{
\item{filepath}{Path to the user's genotype file.}
}
\value{
A tibble with columns: rsid, chromosome, position, and genotype.
}
\description{
Read User Data (e.g., 23andMe genotype file)
Reads the user's genotype file, accommodating headers that may
or may not be commented out with '#'. Keeps rsid, chromosome, position, genotype.
}
